refactor(session): extract standalone helpers to session-helpers.ts (#4246 step 2)
Extract 5 pure functions from session.ts into session-helpers.ts:
- hydrateSessions: rehydrate persisted sessions from schema
- isObjectRecord: type guard for plain objects
- detectModelFromSettings: detect CC model from settings files
- detectIsolationMode: detect worktree isolation mode
- getUiApprovalInput: resolve approval keypress from pane text

No behavioral changes — pure extraction with re-exports.

Changes:
- New: src/session-helpers.ts (113 lines)
- Modified: src/session.ts (1334 → 1258, -76 lines, -6%)
- 5683 tests passed

Part of #4246.
